cpu: aarch64: reorder: separate driver and kernel defs and impls
