+++ src/cpu/aarch64/reorder/jit_uni_reorder.hpp
@@ -19,11 +19,10 @@
 #ifndef CPU_AARCH64_REORDER_JIT_UNI_REORDER_HPP
 #define CPU_AARCH64_REORDER_JIT_UNI_REORDER_HPP
 
-#include <assert.h>
+#include <cassert>
 
 #include "common/c_types_map.hpp"
-#include "common/type_helpers.hpp"
-
+#include "cpu/aarch64/reorder/jit_uni_reorder_kernel.hpp"
 #include "cpu/aarch64/reorder/jit_uni_reorder_utils.hpp"
 #include "cpu/reorder/cpu_reorder_pd.hpp"
 
@@ -31,72 +30,6 @@ namespace dnnl {
 namespace impl {
 namespace cpu {
 namespace aarch64 {
-
-namespace tr {
-struct call_param_t {
-    const void *in = nullptr;
-    void *out = nullptr;
-    const float *src_scales = nullptr;
-    const float *dst_scales = nullptr;
-    int32_t src_zp = 0;
-    int32_t dst_zp = 0;
-    int32_t *compensation_scratch = nullptr;
-};
-
-// The additional structure is needed because
-// using a data structure with tail processing
-// data for non-tail cases reduces kernel
-// performance. This is because there is too
-// much data that has to be transferred to the kernel.
-struct tail_call_param_t {
-    call_param_t base_params;
-    int64_t curr_data_chunks[DNNL_MAX_NDIMS] = {-1};
-    int64_t zeroing_data = static_cast<int64_t>(false);
-    int64_t skip_kernel_execution = static_cast<int64_t>(false);
-};
-
-struct kernel_t {
-    struct desc_t {
-        int id;
-        prb_t prb;
-    };
-
-    kernel_t(const desc_t &desc)
-        : desc_(desc)
-        , compensation_needed_(
-                  desc.prb.req_s8s8_comp || desc.prb.req_asymmetric_comp) {}
-    virtual void operator()(const call_param_t *c) const = 0;
-    virtual void operator()(const tail_call_param_t *c) const = 0;
-    virtual status_t create_kernel() = 0;
-    virtual ~kernel_t() = default;
-
-    /** inits kernel descriptor:
-     *      desc            -- kernel descriptor (output)
-     *      prb             -- transposition problem (input)
-     *      ndims_ker_max   -- limit the maximum number of dimensions kernel
-     *                         will process (optional, 0 -- no limitation) */
-    static status_t desc_init(
-            desc_t &desc, const prb_t &prb, int ndims_ker_max = 0);
-
-    /** creates kernel for the problem described in desc */
-    static kernel_t *create(const desc_t &desc);
-
-    /** Minimal reasonable/desirable kernel size.
-    * The constant might be used to determine how a problem should be split
-    * between kernel and threading driver. */
-    static constexpr size_t ker_prb_size_min = 64;
-
-protected:
-    const desc_t desc_;
-    const prb_t &prb_ = desc_.prb;
-    bool compensation_needed_ = false;
-};
-
-/* TODO: add trans_t class */
-
-struct jit_single_blk_kernel_t;
-
-} // namespace tr
 
 struct jit_uni_reorder_t : public primitive_t {
     using primitive_t::primitive_t;
@@ -161,36 +94,6 @@ private:
 
     const pd_t *pd() const { return (const pd_t *)primitive_t::pd().get(); }
     std::unique_ptr<tr::kernel_t> kernel_;
-};
-
-struct jit_blk_reorder_t : public primitive_t {
-    using primitive_t::primitive_t;
-    struct pd_t : public cpu_reorder_pd_t {
-        using cpu_reorder_pd_t::cpu_reorder_pd_t;
-        DECLARE_COMMON_PD_T("jit:blk", jit_blk_reorder_t);
-
-        tr::prb_t prb_;
-
-    private:
-        static status_t create(reorder_pd_t **reorder_pd, engine_t *engine,
-                const primitive_attr_t *attr, engine_t *src_engine,
-                const memory_desc_t *src_md, engine_t *dst_engine,
-                const memory_desc_t *dst_md);
-
-        // Swap last two nodes, put block 4, 8, 16 nodes to first
-        static void prb_tile_normalize(tr::prb_t &p);
-        friend dnnl::impl::impl_list_item_t;
-    };
-
-    status_t init(engine_t *engine) override;
-    status_t execute(const exec_ctx_t &ctx) const override;
-
-    jit_blk_reorder_t(const pd_t *apd);
-    ~jit_blk_reorder_t() override;
-
-private:
-    const pd_t *pd() const { return (const pd_t *)primitive_t::pd().get(); }
-    std::unique_ptr<tr::jit_single_blk_kernel_t> kernel_;
 };
 
 } // namespace aarch64
